requirements.txt
PyYAML>=6.0
discord.py>=2.3.0
